test(bun): split the runner suite so it stays inside the file-size budget
The subprocess-integration blocks move to their own file rather than raising
the max-lines threshold.
